+++ sdk/dotnet/Ec2/Instance.cs
@@ -511,6 +511,7 @@ namespace Pulumi.Aws.Ec2
         /// <summary>
         /// A list of security group names (EC2-Classic) or IDs (default VPC) to associate with.
         /// </summary>
+        [Obsolete(@"Use of `securityGroups` is discouraged as it does not allow for changes and will force your instance to be replaced if changes are made. To avoid this, use `vpcSecurityGroupIds` which allows for updates.")]
         public InputList<string> SecurityGroups
         {
             get => _securityGroups ?? (_securityGroups = new InputList<string>());
@@ -827,6 +828,7 @@ namespace Pulumi.Aws.Ec2
         /// <summary>
         /// A list of security group names (EC2-Classic) or IDs (default VPC) to associate with.
         /// </summary>
+        [Obsolete(@"Use of `securityGroups` is discouraged as it does not allow for changes and will force your instance to be replaced if changes are made. To avoid this, use `vpcSecurityGroupIds` which allows for updates.")]
         public InputList<string> SecurityGroups
         {
             get => _securityGroups ?? (_securityGroups = new InputList<string>());