+++ sdk/nodejs/ec2/instance.ts
@@ -515,6 +515,8 @@ export interface InstanceState {
     readonly rootBlockDevice?: pulumi.Input<inputs.ec2.InstanceRootBlockDevice>;
     /**
      * A list of security group names (EC2-Classic) or IDs (default VPC) to associate with.
+     * 
+     * @deprecated Use of `securityGroups` is discouraged as it does not allow for changes and will force your instance to be replaced if changes are made. To avoid this, use `vpcSecurityGroupIds` which allows for updates.
      */
     readonly securityGroups?: pulumi.Input<pulumi.Input<string>[]>;
     /**
@@ -667,6 +669,8 @@ export interface InstanceArgs {
     readonly rootBlockDevice?: pulumi.Input<inputs.ec2.InstanceRootBlockDevice>;
     /**
      * A list of security group names (EC2-Classic) or IDs (default VPC) to associate with.
+     * 
+     * @deprecated Use of `securityGroups` is discouraged as it does not allow for changes and will force your instance to be replaced if changes are made. To avoid this, use `vpcSecurityGroupIds` which allows for updates.
      */
     readonly securityGroups?: pulumi.Input<pulumi.Input<string>[]>;
     /**